_restructure/solguard-core/requirements.txt
# SolGuard — Dependencies
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.27.0
python-dotenv>=1.0.0
anthropic>=0.40.0
solana>=0.35.0
solders>=0.21.0
prometheus-client>=0.21.0
